backend/pyproject.toml
[project]
name = "sigint-backend"
version = "0.1.0"
description = "Backend service for SIGINT Wiki - content sync, embeddings, semantic search"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.0.0",
    "httpx>=0.28.0",
    # PyTorch CPU-only to avoid NVIDIA/CUDA dependencies
    "torch>=2.0.0",
    "sentence-transformers>=3.3.0",
    "pgvector>=0.3.0",
    "python-frontmatter>=1.1.0",
    "gitpython>=3.1.0",
    "authlib>=1.3.0",
    "redis>=5.2.0",
    # Authentication dependencies
    "argon2-cffi>=23.1.0",
    "aiosmtplib>=3.0.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
